% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Reading files into R}
\usage{
myread(csv)
}
\arguments{
\item{csv}{}
}
\value{
data frame containing the data
}
\description{
This reads in a csv file from a directory
}
\examples{
dird="\\\\Users\\\\HyDRO-Lab\\\\Desktop\\\\MATH4753\\\\DATAxls\\\\"; myread("DDT.csv")
}
